+++ sdk/communication/Azure.Communication.CallAutomation/src/CallAutomationClientOptions.cs
@@ -16,7 +16,7 @@ namespace Azure.Communication.CallAutomation
         /// <summary>
         /// The latest version of the CallAutomation service.
         /// </summary>
-        internal const ServiceVersion LatestVersion = ServiceVersion.V2023_10_03_Preview;
+        internal const ServiceVersion LatestVersion = ServiceVersion.V2024_06_15_Preview;
 
         internal string ApiVersion { get; }
 
@@ -36,6 +36,8 @@ namespace Azure.Communication.CallAutomation
                 ServiceVersion.V2023_06_15_Preview => "2023-06-15-preview",
                 ServiceVersion.V2023_10_15 => "2023-10-15",
                 ServiceVersion.V2023_10_03_Preview => "2023-10-03-preview",
+                ServiceVersion.V2024_04_15 => "2024-04-15",
+                ServiceVersion.V2024_06_15_Preview => "2024-06-15-preview",
                 _ => throw new ArgumentOutOfRangeException(nameof(version)),
             };
         }
@@ -64,7 +66,17 @@ namespace Azure.Communication.CallAutomation
             /// <summary>
             /// Latest ALPHA3 (1.2.0-alpha) preview of the CallAutomation service.
             /// </summary>
-            V2023_10_03_Preview = 4
+            V2023_10_03_Preview = 4,
+
+            /// <summary>
+            /// Latest GA3 (1.2.0) of the CallAutomation service.
+            /// </summary>
+            V2024_04_15 = 5,
+
+            /// <summary>
+            /// Latest ALPHA3 (1.3.0-alpha) preview of the CallAutomation service.
+            /// </summary>
+            V2024_06_15_Preview = 6
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
     }